pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "i4g"
dynamic = ["version"]
description = "Scratchpad for prototype utilities."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "altair",
    "datasets",
    "faiss-cpu",
    "fastapi",
    "uvicorn[standard]",
    "google-api-python-client",
    "google-auth-httplib2",
    "azure-identity",
    "azure-search-documents>=11.6.0,<12",
    "google-auth-oauthlib",
    "google-cloud-discoveryengine>=0.15.0,<1.0",
    "google-cloud-aiplatform>=1.70.0,<3.0",
    "google-cloud-storage",
    "google-cloud-secret-manager",
    "huggingface-hub>=0.34.0,<1.0",
    "langchain",
    "langchain-chroma",
    "langchain-community",
    "langchain-ollama",
    "ollama",
    "httpx",
    "paddleocr[all]",
    "Pillow",
    "pydantic>=2.6,<3",
    "pydantic-settings>=2.1,<3",
    "python-multipart",
    "google-cloud-firestore",
    "pyodbc",
    "azure-storage-blob",
    "protobuf>=5,<6",
    "pytesseract",
    "python-docx",
    "openpyxl",
    "reportlab",
    "regex",
    "requests",
    "rich",
    "streamlit>=1.51.0,<2",
    "tqdm",
    "trio",
    "alembic>=1.13,<2",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-mock",
    "pytest-anyio",
    "pip-tools",
    "pre-commit",
]

[project.scripts]
i4g-admin = "i4g.cli.admin:main"
run-dataflow = "prefect_gcp.beam.pipeline:main"
i4g-ingest-job = "i4g.worker.jobs.ingest:main"
i4g-report-job = "i4g.worker.jobs.report:main"
i4g-intake-job = "i4g.worker.jobs.intake:main"
i4g-account-job = "i4g.worker.jobs.account_list:main"
i4g-ingest-retry-job = "i4g.worker.jobs.ingest_retry:main"
i4g-dossier-job = "i4g.worker.jobs.dossier_queue:main"

[tool.hatch.version]
path = "VERSION.txt"
pattern = "(?P<version>.+)"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (slower)",
]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
